Updated mosaic_tiles function nodata handling
There was an issue when the input tiles had different nodata values because the function was using the nodata value from the first tile as the mosaic nodata value.  The function now gets the nodata for each tile separately and converts nodata cells to the mosaic nodata value before adding to the mosaic raster.

Also added the resampling type as an input to the function so it can be passed through to project raster.

Other miscellaneous formatting and doc string changes.